+++ lib/color_filter/color_filter_component.dart
@@ -1,4 +1,5 @@
 import 'package:bonfire/base/game_component.dart';
+import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/color_filter/game_color_filter.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:flutter/widgets.dart';
@@ -57,7 +58,7 @@ class ColorFilterComponent extends GameComponent
       end: color,
     );
 
-    gameRef.getValueGenerator(
+    generateValues(
       duration,
       onChange: (value) {
         config.color = _tween?.transform(value);
@@ -67,6 +68,6 @@ class ColorFilterComponent extends GameComponent
         onFinish?.call();
       },
       curve: curve,
-    ).start();
+    );
   }
 }